requirements.txt
Flask==2.1.0
gunicorn
Flask-SocketIO
eventlet==0.30.2
gevent-websocket
python-dotenv
pymongo
opencv-python
requests
numpy
